# server/src/interfaces/http/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a socco project",
    "version": "1.0.0",
    "title": "SOCCO Backend",
    "termsOfService": "",
    "contact": {
      "email": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
  "paths": {
...
    "/store/order": {
      "post": {
        "tags": [
          "store"
        ],
        "summary": "Place an order for a products",
        "description": "",
        "operationId": "placeOrder",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "order placed for purchasing the product",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Cart"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Invalid Order"
          }
        }
      },
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Get orders for customer",
        "description": "",
        "operationId": "retrieveOrder",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Orders"
              }
            }
          },
          "400": {
            "description": "Invalid Order"
          }
        }
      }
    },
    "/store/order/{orderId}": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Find purchase order by ID",
        "description": "For valid response try integer IDs with value >= 1 and <= 10. Other values will generated exceptions",
        "operationId": "getOrderById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of order that needs to be fetched",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "put": {
        "tags": [
          "store"
        ],
        "summary": "Find purchase order by ID",
        "description": "For valid response try integer IDs with value >= 1 and <= 10. Other values will generated exceptions",
        "operationId": "updateOrderById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "ID of order that needs to be fetched",
            "required": true,
            "type": "integer",
            "maximum": 10,
            "minimum": 1,
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Update order with status",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
...
    "/store/cart": {
      "post": {
        "tags": [
          "cart"
        ],
        "summary": "Add item to cart with default quantity as 1",
        "description": "",
        "operationId": "addToCart",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Add product in the cart",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Cart"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Cart"
            }
          },
          "400": {
            "description": "Invalid Cart"
          }
        }
      },
      "get": {
        "tags": [
          "cart"
        ],
        "summary": "Get items from cart",
        "description": "",
        "operationId": "getCart",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Cart"
              }
            }
          },
          "400": {
            "description": "Invalid Cart"
          }
        }
      },
      "put": {
        "tags": [
          "cart"
        ],
        "summary": "Update item to cart with quantity",
        "description": "",
        "operationId": "updateCart",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Update product in the cart",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CartItemQuantity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Cart"
            }
          },
          "400": {
            "description": "Invalid Cart"
          }
        }
      },
      "delete": {
        "tags": [
          "cart"
        ],
        "summary": "Delete product from cart",
        "description": "",
        "operationId": "removeCartItem",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Delete product in the cart",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RemoveCartItem"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Cart"
            }
          },
          "400": {
            "description": "Invalid Cart"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Order": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "integer",
          "format": "int64"
        },
        "customerId": {
          "type": "integer",
          "format": "int64"
        },
        "communityId": {
          "type": "integer",
          "format": "int64"
        },
        "totalAmount": {
          "type": "integer",
          "format": "int32"
        },
        "shipDate": {
          "type": "string",
          "format": "date-time"
        },
        "createDate": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "comments": {
          "type": "string"
        },
        "authCode": {
          "type": "string"
        },
        "reference": {
          "type": "string"
        },
        "shippingId": {
          "type": "integer"
        },
        "taxId": {
          "type": "integer"
        }
      }
    },
    "Orders": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "integer",
          "format": "int64"
        },
        "productId": {
          "type": "integer",
          "format": "int64"
        },
        "productName": {
          "type": "string"
        },
        "quantity": {
          "type": "integer"
        },
        "thumbnail": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "enum": ["ORDER_PLACED", "PROCESSING", "SHIPPED", "OUT_FOR_DELIVERY", "DELIVERED", "CANCELLED"]
        },
        "community_id": {
          "type": "integer",
          "format": "int64"
        },
        "community_name": {
          "type": "string"
        },
        "agent_id": {
          "type": "integer",
          "format": "int64"
        },
        "agent_name": {
          "type": "string"
        },
        "agent_phone": {
          "type": "string"
        }
      }
    },
...
    "OrderStatus": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32",
          "example": "Status 1 for success payment, 2 for cancel payment"
        },
        "comments": {
          "type": "string"
        },
        "authCode": {
          "type": "string",
          "example": "PaymentId recieved from the payment gateway"
        },
        "reference": {
          "type": "string",
          "example": "Json Object received from the payment gateway"
        }
      }
    },
    "Cart": {
      "type": "object",
      "properties": {
        "itemId": {
          "type": "integer",
          "format": "int64"
        },
        "cartId": {
          "type": "integer",
          "format": "int64"
        },
        "productId": {
          "type": "integer",
          "format": "int64"
        },
        "communityId": {
          "type": "integer",
          "format": "int64"
        },
        "attributes": {
          "type": "string"
        },
        "quantity": {
          "type": "integer"
        },
        "buyNow": {
          "type": "integer"
        },
        "createdDate": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "CartItemQuantity": {
      "type": "object",
      "properties": {
        "quantity": {
          "type": "integer"
        },
        "productId": {
          "type": "integer"
        }
      }
    },
    "RemoveCartItem": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "integer"
        }
      }
    },
...
  }
}